create thread_manager program
